Chat_With_Document/requirements.txt
unstructured
langchain-unstructured
tiktoken
pinecone
pypdf
groq
langchain
langchain-community
langchain-pinecone
langchain-text-splitters
langchain-groq
streamlit
python-dotenv
sentence-transformers
spacy

